[CodeGen] Add runtime check when using MemTagSanitizer
Currently, on Darwin platforms, it is possible to compile a program with MemTagSanitizer and run it on hardware that does not support MTE. In such a case generally you will reach an illegal instruction, which can be confusing.

This patch adds a global constructor that first checks that you are running on supported hardware, and exits with an error message indicating the issue if not.

rdar://162615121